Update cmake ARCHIVE_DIR paths for Open XL
With disabling zos_rename_exports.cmake for Open XL, the sidedeck
files should be generated in the same directory as the shared libs.
As such the "/lib" suffix for the archive directory should not
be needed.

Additional changes will be required to the extensions repos in order
to support these changes. (e.g. updates to expected paths of
side-decks for Copy.gmk)

Signed-off-by: Gaurav Chaudhari <Gaurav.Chaudhari@ibm.com>