Handle decompile paths when mounting a vthread
When decompileing a blocked continuation, the return address will be
patched to point to the decompile helper when we attempt to enter the
continuation. This helper assumes the JIT resolve frame is still on the
stack. Instead of popping the resolve frame just jump to the return
address and leave the frame in place. The decompile helper will pop the
frame.

Signed-off-by: tajila <atobia@ca.ibm.com>