Cmdline options for CS taxation headroom
Private command line options to tune taxation headroom

Valid values to provide are any positive integer, but the map to floats
by subtracting 100 (possibly leading to negative values) and dividing by
100.

The current internal defaults are -0.1 to +0.2, what would map into 90
and 120 for the command line option values.

Values 0 and 0 should be provided to disable the taxation.