Set known object index for condy and method type table entry symrefs
These symbols satisfy isFixedObjectRef(), allowing in some circumstances
for the JIT to attempt to load and dereference them at compile time.
Soon the compiler will start to keep track within each compilation of
where each known object was found, and having known object indices for
these symrefs will facilitate the tracking. The dereferencing paths also
handle known objects. They need to be aware of isFixedObjectRef() only
because there are some for which the known object index is missing. So
setting known object indices on these symrefs as appropriate will allow
that code to be simplified (as a follow-up).