Recognize new converter method
This commit recognizes a new converter method in
IBM Semeru Runtime for z/OS version 11 and newer -
SingleByte$Encoder_encodeFromLatin1Impl. The new
method is analagous to an existing IBM SDK 8 converter
method so the previous acceleration code in Value
Propagation will be used to accelerate the new
method as well.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>