Do not disclaim on tmpfs/ramfs/nfs
OpenJ9 JVM can disclaim memory used for infrequently used data
to a backing file (found on /tmp) or on swap.
This commit prohibits the disclaiming from taking place if
the filesystem for the /tmp directory is tmpfs/ramfs (which
use RAM) or nfs (which has an upredictable access latency).
In such cases the disclaiming will be performed on swap if
swap is available and the user allows it (i.e. if
-Xjit:dontDisclaimMemoryOnSwap was not specified). If that
is not the case, then the disclaim mechanism will be completely
disabled.

Signed-off-by: Marius <mpirvu@ca.ibm.com>