Pass correct value of isUnsafeGet from createUnsafeGetWithOffset()
...to genCodeForUnsafeGetPut(). In particular, pass isUnsafeGet=true.

The incorrect value has been causing createAnchorNodesForUnsafeGetPut()
to create compressed refs anchors as though for an unsafe put, i.e.
wrapping the entire tree (including the top-level node) in an anchor,
and replacing the original tree. This is correct for unsafe put, since
the original top-level node is the indirect store to unsafe shadow. But
for unsafe get, it would produce something like the following:

    compressedRefs
      astore <temp slot $result>
        aloadi <unsafe shadow sym>
          aladd
            aload <temp slot $obj>
            lload <temp slot $offset>
      lconst 0

where the compressedRefs anchor unexpectedly contains a direct store.
Such a tree is liable to cause end-to-end miscompilation. As an example,
escape analysis has been observed to remove the temporary store, leaving
the result of the unsafe load unused.

Passing the correct value allows the compressedRefs anchor and the
temporary store to be generated correctly as separate trees, e.g.:

    compressedRefs
      aloadi <unsafe shadow sym>
        aladd
          aload  <temp slot $obj>
          lload  <temp slot $offset>
      lconst 0
    astore <temp slot $result>
      ==>aloadi <unsafe shadow sym>

We used to pass true, but it appears that it was inadvertently changed
to false in c79d879a041f08d10b6ade76035c7fb4b8134122.