Update J9SP register on interp exit
Upon exit from the interpreter on aarch64, the SP (x20) register is
re-loaded from the ELS. This would be the SP upon calling the
jitMonitor* helper. If there is a decompiation the vthread will resume
with the jitResolveFrame on the stack. This means the SP is incorrect.
This PR reloads the latest SP upon exiting the interpreter, this will
account for any frames pushed between the point the JIT calls the helper
and when the vthread returns after blocking.

Fixes https://github.com/eclipse-openj9/openj9/issues/22206

Signed-off-by: tajila <atobia@ca.ibm.com>