RAM Class Persistence: Fix class not found
When modules are used in applications, the application class loader
is allocated and causes applications to fail. The problem is that
the java object of builtin application class loader is constructed,
but not initialized. The fix is to initialize the application class
loader upon construction like boot and platform class loaders.

Fixes: #22287