RAM Class Persistence: Fix class not found
When modules are used in applications, the application class loader
is allocated and causes applications to fail. The problem is that
the java object of builtin application class loader is constructed,
but not initialized. The fix is to initialize the application class
loader upon construction like boot and platform class loaders.
The changes is to implement JNI interface VM.isRestoreRun() and
VM.rcpAssignClassLoader() to initialize the builtin class loaders
by OutOfLineINL_com_ibm_oti_vm_VM_rcpAssignClassLoader() for RCP
restore run. A new constant J9_CLASSLOADER_TYPE_APP is added to
identify the application class loader.

Fixes: #22287

Co-authored-by: Tobi Ajila <tobi_ajila@ca.ibm.com>
Co-authored-by: Babneet Singh <sbabneet@ca.ibm.com>