Detect any monitorenter before point of dynamic loop transfers
Dynamic loop transfer can transfer control from the interpreter into a
JIT-compiled method after the point at which monitorenter bytecode
instructions have been executed.  With Java 24 and later, the JVM uses
the JIT's live monitor metadata to locate monitors that might need to be
detached from carrier threads when the owning virtual thread is
unmounted.  This is enabled if the
J9_EXTENDED_RUNTIME3_YIELD_PINNED_CONTINUATION flag is set in
extendedRuntimeFlags3, which can be turned off by specifying the option
-XX:-YieldPinnedVirtualThreads.

If the JIT is unable to determine that no monitorenters will be executed
prior to the entry point for dynamic loop transfer, it must fail the
compilation.  This change follows predecessor paths in the CFG from the
blocks identified as potential DLT entry points to see whether a
monitorenter could have been encountered before the DLT occurred.  If
that might be possible, the compilation fails.

Also, if the JVM requires correct live monitor maps, force the
disableLiveMonitorMetadata JIT option off.

Finally, the disableInlineMonEnt and disableInlineMonExit options can be
set in a fashion that's specific to a particular method compilation.
Whether they should be forced off needs to be considered earlier in
J9::Options::feLatePostProcess.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>