Add spill registers in walkTransitionFrame()
There are cases that the stack walker walk jit frames in transition
frame e.g. when a virtual thread fails to enter a synchronized method.
We still walk the register map in transition frame if there are object
slots in registers but we don't update the register table with spilled
registers which is causing failure.
This PR add procedure to update register table with spilled register
when it is in method prologue for synchronized method cases.

Closes: #22079
Backport of: #22303
Signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>