Adjust sampling thread rate based on CPU utilization
Previously, the rate of the JIT sampling thread was decided
only based on the number of "active" Java threads (threads
with VM access) using a very simple formula:
`period = max(10 ms, 2ms * numActiveThreads)`. The idea was
to limit the overhead of sampling when the number of threads
increases.
However, when the JVM is heavily starved of CPU resources,
the sampling thread will still beat every 10 ms, while the
application thread may get little time on the CPU. This is
another example of sampling thread vs application thread
imbalance which this commit attempts to fix.
Starting with this commit, the rate of the sampling thread
will also be adjusted based on the CPU utilization of the
JVM and the entire machine. Basically, we want the sampling
thread to be less active when we detect that the JVM is
CPU constrained (i.e. the JVM does not use its entire CPU
entitlement and the machine has no idle cycles).
This commit also implements a small change related to the
amount of load visible to the IProfiler which will allow
the usage of the dedicated IProfiler thread even in
single CPU environments. This seemed to improve startup
time by a small margin.

Signed-off-by: Marius <mpirvu@ca.ibm.com>