Use AArch64 IPC register for indirect branches in JIT runtime
The macro `BRANCH_VIA_VMTHREAD` in the JIT runtime uses the `x8` register to
perform an indirect branch to an arbitrary function. However, this macro may be
used in contexts where it is not safe to clobber `x8`.

Use the conventional AArch64 ABI register `x16` for this purpose instead. The
JIT does not register assign `x16` so there are no conflicts.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>