Z: Allow helper populate iPIC slots
We implemented last iTable cache and iTable lookup in jit. Now we dipatch
to the method implementation in jit and it prevent the helper to populate
iPIC slots.
This PR check the value of last iPIC slot method and branch to OOL iTable
cache when there is a non zero value there.

signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>