Make client and server agree on exact permanent loaders
...within each compilation.

There will be at least two data structures that are sensitive to the
set of permanent loaders. These could potentially allow the precise set
of permanent loaders to vary between the client and server, but the
reasoning needed to determine whether certain kinds of variation are
allowed gets hairy very quickly. It's much simpler if there is only one
set of permanent loaders that will be used for all purposes on both the
client and the server.

This change also simplifies reasoning about which permanent loaders the
server already has. The client no longer attempts to keep track of that
at all. Instead, that information is purely in the purview of the
server, where it's easy to maintain consistency.