Use comp->permanentLoaders() for client retained methods analysis
This was avoiding getting the permanent loaders from the compilation
object to guard against a scenario where the server side of the
compilation was aware of more permanent loaders than the client side.

That scenario is no longer possible because a compilation on the server
now limits itself to the same set of permanent loaders that the
corresponding compilation is aware of on the client. It's no longer
necessary for J9::RetainedMethodSet to allow the caller to specify the
permanent loaders on creation.