Notify unblocker during OMR wait
There are currently two paths where the unblocker may spin in a tight
loop. The first is if the continuation list is empty and the second is
if a virtualThread is waiting on a lock owned by a platform thread. Both
of these cases are handled by this PR.

The first case is handled by removing the
`(NULL == vm->blockedContinuations)` check at the start of the function.
The second case is handled by removing the hasPlatformThreadWaiting code
and instead notifying the unblocker anytime a platform thread waits on a
monitor.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>