MH transformer: compute known object index for all nodes in postorder
...instead of waiting for getObjectInfoOfNode() to request it.

This allows method handle transformer's handling of autos to be correct
for arbitrary IL, including the following:

          astore tmp1
            ... (unknown)
          treetop
    n1n     aload tmp1
          astore tmp1
            ... (obj1)
          astore tmp2
    n1n     ==>aload tmp1

In this case, previously MH transformer would not determine the known
object index for n1n until it was requested at the store to tmp2, at
which point it would incorrectly conclude that n1n must be obj1. By
computing the known object index for every address-typed node at a
well-defined time during the traversal in visitNode(), MH transformer
will now correctly see in this example that n1n (and therefore tmp2)
should not have a known object index.

This pattern of trees is not currently expected to be encountered in MH
transformer, since it only runs during ilgen opts, and only for
LambdaForm-generated methods, for which the bytecode is highly
constrained.

Additionally:

- Check up front for the presence of the known object table so that it's
  not necessary to check repeatedly.

- Consolidate the setting of the per-node known object index into just a
  single location. Later, for const refs, this consolidation will make
  it easy to do something else instead (transmute the node into a const
  ref load).

- Add braces around single-statement bodies of control flow structures
  in edited methods.