Use early exits in MH transformer checkVarHandleGenericType() handler
This allows the bulk of the body to be unindented by two levels, and it
creates a natural location for an additional early exit that will be
added later for const refs. It's also consistent with the existing early
exit due to the environment variable TR_disableFoldingVHGenType.

Additionally, replace the trace message with performTransformation(),
and add braces around single-statement bodies of control flow structures
in this handler.