Add const provenance edges where necessary
No edges are added in code that is used only for J9VM_OPT_METHOD_HANDLE,
e.g. runFEMacro. Const refs will require J9VM_OPT_OPENJDK_METHODHANDLE.

With JITServer, it doesn't matter whether an edge is added on the server
or on the client. When it comes time to search the const provenance
graph near the end of the compilation, the client will send all of its
edges to the server, which will incorporate them in the search. This
communication is not implemented in this commit, but will be implemented
along with the search itself in a later commit.