Delete refineInvokeCacheElementSymRefWithKnownObjectIndex()
The only caller will now instead get the known object index from the
existing query getKnotIndexOfInvokeCacheArrayAppendixElement().

This leaves improving the symref to the caller, which will be helpful
later for const refs. With const refs, we'll no longer generate the same
kind of improved symref, but we'll still need the known object index.