Always build a new cuda image
Always build a new cuda image to ensure we are
using the latest base image.

Also change the name to include the tag in the base
name. This is to handle the base name being the same
for centos6,7 which are now only distinguishable
based on the tag.

Remove image as an arg in the Dockerfile. Insead,
hardocde it in the FROM line. Eliminates a build warning.

Change case of AS in cuda line to eliminate build warning.

Resolves #22434
Related #20154

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>