JITServer shared fanin cache bug fix
This commit fixes a bug related to the JITServer shared fanin cache:
if the server determines that the quality of the fanin data from
the shared profile cache is better than the quality of the fanin
data sent by the client, it calls `loadFaninDataFromSharedProfileCache()`
If this fails for whatever reason, the server will use the fanin data
from the client. Because of a bug, the data from the shared profile
cache was never returned. This commit fixes this bug so that the
server can effectively use the fanin data from shared profile cache.

Signed-off-by: Marius <mpirvu@ca.ibm.com>