Support for dumping all IProfiler data to a file
When `TR_DumpIProfilerData` environment variable is
defined, all the interpreter profiler data will be dumped
to a file at JVM shutdown time. The file is named:
ipdata and will be suffixed with the date and time of
the creation time plus the PID of the JVM:
ipdata.DATE.TIME.PID

The file is in readable text format and the IProfiler data
is grouped by method. Here is an example of how the data
may look like

Method: com/ibm/ws/container/service/metadata/internal/MetaDataManager.getMetaDataImpl();
	Offset 1	branch samples=6	 taken=0
	Offset 15	JBcheckcast samples=6 TooBig=0
		W:   3	M:0xb65200	com/ibm/ws/webcontainer/osgi/metadata/WebComponentMetaDataImpl
		W:   1	M:0xb64d00	com/ibm/ws/webcontainer/osgi/metadata/WebCollaboratorComponentMetaDataImpl
		W:   1	M:0x9f7600	com/ibm/ws/webcontainer/osgi/metadata/WebModuleMetaDataImpl
		W:   1	M:0xffffffff	OTHERBUCKET
        Offset 20       switch samples=10

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>