Modify memory option rounding to avoid overflow of 64-bit signed values
Memory option values are rounded up to be pointer-width aligned unless this
leads to overflow in either the native or 64-bit signed representation. In
these cases we truncate (round down) to the next pointer-width aligned
value.

At the same time we add a bounds check on the MaxDirectMemorySize value to
ensure it maps correctly to a Java long before passing in to the JVM
initialization.