Modify memory option rounding to avoid overflow
Memory option values are rounded up to be pointer-width aligned unless
this would lead to a value greater than U_32_MAX or I_64_MAX on 32 and
64 bit platforms respectively. In these cases we truncate (round down)
to the next pointer-width aligned value.

At the same time we add a bounds check on the MaxDirectMemorySize value
on 64 bit platforms to ensure it maps safely to a Java long before
passing in to the JVM initialization.

Fixes: #22480
Signed-off-by: Chris Dennis <chris.w.dennis@ibm.com>