x86: Fix interface dispatch performance
X86PicBuilder expects that the last instruction before the
done label is a JNE instruction to the lookup routine. The
JNE should not jump to the lookup routine, but is required
because this code may be patched. This commit introduces a
JNE instruction to replace a JMP instruction, which always
call the lookup routine, which incurs a performance
penalty.

Signed-off-by: Bradley Wood <bradley.wood@ibm.com>