Replace CpuUtilization->isFunctional() with hasValidData()
Existing code uses CpuUtilization info after a test on
CpuUtilization->isFunctional(). However, to report CPU
utilization we need two data collections at different points
in time (the JVM computes the difference of two absolute values).
Until the two data collections are available, the data reported
by CpuUtilization cannot be trusted (for instance the JVM CPU
utilization shows up as 0).

This commit introduces a new getter, CpuUtilization->hasValidData()
that returns true if at least two CPU data collections have been performed.
The existing code has been changed to replace isFunctional() with
hasValidData() for most of the queries (is Functional() is still used
when updating the CpuUtilization information).

To minimize the window of time when the CpuUtilization is functional
but does not have valid data, the sampling thread will attempt to
collect the second CPU data point as soon as possible.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>