Add support for disclaiming class memory
Add support for disclaiming class memory. This will be done in stages.

This PR disclaims only a portion of RAM class memory (sub4g). In the
future the remainder of the RAM and ROM class will be disclaimed. This
PR allows the option specify if both ROM and RAM will be disclaimed or
just one of them. This is useful for testing purposes. There is no
intention to document or  publicize any of the options at this point,
this can be done after some extensive testing.

When disclaim is enabled, sub4g memory is backed by an anonyous
tempfile. Currently, only the JIT will trigger class memory disclaim.
This will be done periodically. In the future, disclaim may also be done
at checkpoint time.

Signed-off-by: tajila <atobia@ca.ibm.com>