[JDK25] Fix StructuredTaskScope/StructuredThreadDumpTest
Make the minimal changes needed for StructuredThreadDumpTest to pass,
without fully implementing jdk.internal.vm.ThreadSnapshot.

Reorder operations so objects do not require a special frame for GC
protection when VM access is not held. This also resolves a crash.

Remove setting of jdk.internal.vm.ThreadSnapshot.blockerObject to avoid
a crash in jdk.internal.vm.ThreadDumper.

A full implementation of ThreadSnapshot is not required for release,
since it is an internal API used mainly for introspection in test code.
It can be developed gradually as further tests require support.

Related: #22220

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>