Set register dep to exclude GPR0 for temp2Reg in inlineCompareAndSet
The temp2Reg is used by `VMnonNullSrcWrtBarCardCheckEvaluator` in some
cases in a `stbx` instruction, which can't have GPR0 as the RA argument
or it will act upon it as a null value.
This causes the an unexpected register spill in the middle of an
internal control-flow sequence that might not be fully executed.