Adjust sub-4G RAM class segment to hold 64KB worth of J9Class headers
Currently, sub-4G RAM class segments are limited to 32 J9Class headers.
This change updates the allocation to store as many J9Class headers as
fit in a 64KB segment.

The goal is to reduce the total number of sub-4G segments required, which
should simplify memory management and may also help mitigate a startup
regression observed in recent builds. It will also benefit platforms such
as PPC, where the default page size is 64KB, by improving the ability to
disclaim memory when segments are released.

Macro RAM_CLASS_HEADERS_PER_SEGMENT is used to compute the capacity, so
the number of headers adjusts automatically based on sizeof(J9Class).

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>