Switch to interpreter on recomp failure if the JIT body is invalidated
...for any reason whatsoever.

In particular, don't check all possible invalidation reasons separately
in the condition that decides whether or not to switch to interpreter in
methodCannotBeRecompiled(). Listing them out this way duplicates more
logic than necessary, and it makes it easy to miss one, e.g. when a new
reason is added.

In fact, one new use of JIT body invalidation was not accounted for in
these conditions: the body could have been invalidated because an
inlined method was unloaded.

As a result, it was possible for methodCannotBeRecompiled() to take note
that recompilation had failed (via setHasFailedRecompilation()) without
further patching the prologue, which could cause execution to enter an
infinite loop where:
1. the invalidated body's prologue tries to recompile synchronously, but
2. the JIT sees that recompilation has already finished (failed), so
3. it returns the existing start PC (for the invalidated body), and
4. execution jumps to the invalidated prologue again, returning to (1).

Some additional potential consequences of the old conditions (from
inspection, unverified):

- methodCannotBeRecompiled() might sometimes have restored the original
  prologue, allowing the JIT body to be re-entered even though it was
  invalidated (after which is it never supposed to be entered again).

- Execution could have been unnecessarily sent to the interpreter if a
  recompilation failed for which the original JIT body used preexistence
  but had not been invalidated.