Z: Fix multianewarray evaluator register dependency bug
Although the number of dimensions in multianewarray node is constant
and not used via a register in the evaluator's mainline ICF path, it is
referenced in the out-of-line (OOL) section for setting up call
parameters. If the dimension node has multiple reference counts and is
not explicitly tied to the mainline dependency, it may be spilled in
the OOL path and reverse-spilled elsewhere, potentially leading to
unexpected behavior. This change ensures that the dimension node is
evaluated and attached to the mainline dependency to maintain
correctness and prevent such issues.

Fix: #22757
signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>