Initialize temporary used to test whether heapification is needed
The heapifyForColdBlocks method of Escape Analysis creates a local
temporary that it uses to hold a reference to a stack allocated object
at the point of the stack allocation.  If the object might need to be
heapified, it generates IL that tests whether that temporary still holds
a reference to the stack allocated object, and if it does, it performs
the heapification and updates the temporary.  However, the temporary was
not explicitly initialized on entry to the method.

During Code Generation J9::CodeGenerator::createStackAtlas will ensure
that any autos that might contain references on entry to the method are
zero initialized.  Despite that, it's better practice to ensure that
such initialization is explicitly performed explicitly in the IL, as it
allows, for instance, other analyses to rely on all the definitions seen
in the IL for a particular auto as enumerating the possible values the
auto can hold at particular points without having to consider the
possibilty that the auto might hold some as yet to be determined default
value that it will acquire during Code Generation.

In the spirit of upholding that better practice, this changes
heapifyForColdBlocks to generate IL at method entry that zero
initializes the temporary variable.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>