Override numInterfacesImplemented() fe query at the JITServer
`TR_J9VMBase::numInterfacesImplemented(J9Class *clazz)` is a frontend
query that returns the number of interfaces implemented by a j9class.
The implementation retrives the iTable of the class and iterates through
the interfaces in the iTable. Each such iteration means a message sent
by the server to the client.
To reduce the number of messages exchanged by the client and server,
this commit overrides the implementation of `numInterfacesImplemented()`
at the server. The new implementation will have the server send a single
message with the desired j9class. The client computes the number of
interfaces implemented by that j9class and returns the answer to the
server.

Issue: #22780

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>