Enable MethodHandleTransformer opt for peeking ILGen methods
Peeking ILGen is currently mainly performed for MethodHandle or
VarHandle scenarios to propagate object info along call graphs.
However, in cases where, for example, a call node corresponding to
an invokehandle/invokedynamic bytecode within a non-root method
loads from an auto slot containing a MethodHandle/VarHandle object,
the object info would not be available even with peeking ILGen.
This would result in failure to inline along the
invokehandle/invokedynamic operation. MethodHandleTransformer
is part of ilgenStrategyOpts that would only be triggered for
LambdaForm generated methods, and this change will expand the role
of this opt to also be triggered for peeking ILGen.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>