Remove useless/problematic code from sun.misc.Unsafe.registerNatives()
In Java 25+, jdk.internal.misc.Unsafe.INVALID_FIELD_OFFSET has type
long, not int, so a NoSuchFieldError is thrown by GetStaticFieldID().
This occurs before the VM is capable of properly handling exception
dump filters (internally, the offset of Throwable.walkback has not
been captured for use by J9VMJAVALANGTHROWABLE_WALKBACK(), etc.).

If the field exists, it is assigned the value -1, but, in all Java
versions, the field would already have that value. This removes the
code that is problematic in Java 25+ and just useless in general.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>