Reserve some postponed hooks so they can be registered later
Some events are postponed because the hook interface is not available,
but others can be reserved so registration can succeed later.

Detect any failure of J9HookRegisterWithCallSite, not just the last
call. Add a tracepoint to capture more detail on failure.

One-shot events are disabled after they occur and cannot be re-enabled.
During a CRIU snapshot or after a CRIU restore, attempting to register
or reserve those events will fail - such failures must be ignored.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>